refactor: split external ticket and hidden external ticket
